Adding Gallery Component.
