+++ public/data/events.json
@@ -10,6 +10,12 @@
     "images": [
       "https://via.placeholder.com/500?text=Image+1",
       "https://via.placeholder.com/500?text=Image+2",
+      "https://via.placeholder.com/500?text=Image+1",
+      "https://via.placeholder.com/500?text=Image+2",
+      "https://via.placeholder.com/500?text=Image+1",
+      "https://via.placeholder.com/500?text=Image+2",
+      "https://via.placeholder.com/500?text=Image+1",
+      "https://via.placeholder.com/500?text=Image+2",
       "https://via.placeholder.com/500?text=Image+3",
       "https://via.placeholder.com/500?text=Image+4"
     ],
